create test for transaction

+++ src/main/java/com/example/jakartauni/rest/CurrencyResource.java
@@ -5,13 +5,17 @@ import com.example.jakartauni.currency.CurrencyService;
 import com.example.jakartauni.rest.dto.ResponseMessageDto;
 import jakarta.ejb.EJB;
 import jakarta.ws.rs.*;
+import jakarta.ws.rs.core.Context;
 import jakarta.ws.rs.core.MediaType;
 import jakarta.ws.rs.core.Response;
+import jakarta.ws.rs.core.SecurityContext;
 
 @Path("/currency")
 public class CurrencyResource {
     @EJB
     private CurrencyService currencyService;
+    @Context
+    private SecurityContext securityContext;
 
     @GET
     @Path("/{currencyId}")
@@ -24,9 +28,22 @@ public class CurrencyResource {
     }
 
     @POST
+    @Path("/transaction-test")
+    @Consumes(MediaType.APPLICATION_JSON)
+    @Produces(MediaType.APPLICATION_JSON)
+    public Response transactionTest(Currency currency) {
+        currencyService.transactionTest(currency);
+        return Response.status(Response.Status.CREATED).build();
+    }
+
+    @POST
     @Consumes(MediaType.APPLICATION_JSON)
     @Produces(MediaType.APPLICATION_JSON)
     public Response addCurrency(Currency currency) {
+        boolean isAdmin = securityContext.getUserPrincipal().getName().equals("admin");
+        if (!isAdmin) {
+            return Response.status(Response.Status.UNAUTHORIZED).build();
+        }
         if (currencyService.findCurrencyByName(currency.getName()).isPresent()) {
             String message = String.format("Currency with name [%s] already exists.",
                     currency.getAbbreviation());
@@ -53,6 +70,10 @@ public class CurrencyResource {
     @Consumes(MediaType.APPLICATION_JSON)
     @Produces(MediaType.APPLICATION_JSON)
     public Response updateCurrency(@PathParam("id") Long id, Currency updatedCurrency) {
+        boolean isAdmin = securityContext.getUserPrincipal().getName().equals("admin");
+        if (!isAdmin) {
+            return Response.status(Response.Status.UNAUTHORIZED).build();
+        }
         if (updatedCurrency.getName() == null || updatedCurrency.getAbbreviation() == null) {
             String message = "Currency should have name and abbreviation";
             ResponseMessageDto messageDto = ResponseMessageDto.builder().message(message).build();
@@ -73,6 +94,10 @@ public class CurrencyResource {
     @Path("/{id}")
     @Produces(MediaType.APPLICATION_JSON)
     public Response deleteCurrency(@PathParam("id") Long id) {
+        boolean isAdmin = securityContext.getUserPrincipal().getName().equals("admin");
+        if (!isAdmin) {
+            return Response.status(Response.Status.UNAUTHORIZED).build();
+        }
         return currencyService.findCurrencyById(id)
                 .map(existingCurrency -> {
                     currencyService.deleteCurrency(existingCurrency);
